some new changes in app.jsx
