+++ particles.config.js
@@ -1,75 +1,72 @@
-// const particlesConfig =
-// {
-//   autoPlay: true,
-//   background: {
-//     color: {
-//       value: '#0d47a1',
-//     },
-//   },
-//   fpsLimit: 120,
-//   interactivity: {
-//     events: {
-//       // onClick: {
-//       //   enable: true,
-//       //   mode: 'push',
-//       // },
-//       onHover: {
-//         enable: true,
-//         mode: 'repulse',
-//       },
-//       resize: true,
-//     },
-//     modes: {
-//       push: {
-//         quantity: 4,
-//       },
-//       repulse: {
-//         distance: 100,
-//         duration: 1.4,
-//       },
-//     },
-//   },
-//   particles: {
-//     color: {
-//       value: '#ffffff',
-//     },
-//     links: {
-//       // color: '#ffffff',
-//       distance: 150,
-//       enable: true,
-//       opacity: 0.5,
-//       width: 1,
-//     },
-//     move: {
-//       direction: 'none',
-//       enable: true,
-//       outModes: {
-//         default: 'bounce',
-//       },
-//       random: false,
-//       speed: 3,
-//       straight: false,
-//     },
-//     number: {
-//       density: {
-//         enable: true,
-//         area: 800,
-//       },
-//       value: 80,
-//     },
-//     opacity: {
-//       value: 0.5,
-//     },
-//     shape: {
-//       type: 'circle',
-//     },
-//     size: {
-//       value: { min: 1, max: 5 },
-//     },
-//   },
-//   detectRetina: true,
-// }
-
-
-
-// export default particlesConfig;
+const particlesConfig = {
+  background: {
+    color: {
+      value: 0,
+    },
+    // opacity: 1,
+  },
+  // clear: true,
+  fpsLimit: 120,
+  interactivity: {
+    events: {
+      onClick: {
+        enable: true,
+        mode: 'push',
+      },
+      onHover: {
+        enable: true,
+        mode: 'repulse',
+      },
+      resize: true,
+    },
+    modes: {
+      push: {
+        quantity: 4,
+      },
+      repulse: {
+        distance: 100,
+        duration: 1.4,
+      },
+    },
+  },
+  particles: {
+    color: {
+      value: '#ffffff',
+    },
+    links: {
+      // color: '#ffffff',
+      distance: 150,
+      enable: true,
+      opacity: 1.5,
+      width: 1,
+    },
+    move: {
+      direction: 'none',
+      enable: true,
+      outModes: {
+        default: 'bounce',
+      },
+      random: false,
+      speed: 5,
+      straight: false,
+    },
+    number: {
+      density: {
+        enable: true,
+        area: 800,
+      },
+      value: 80,
+    },
+    opacity: {
+      value: 0.5,
+    },
+    shape: {
+      type: 'circle',
+    },
+    size: {
+      value: { min: 1, max: 5 },
+    },
+  },
+  detectRetina: true,
+};
+export default particlesConfig;
