+++ src/components/ParticlesBackground.jsx
@@ -1,31 +1,33 @@
-// // import React from 'react';
-// import particlesConfig from '../../config/particles.config';
-// import { useCallback, useEffect, useState } from 'react';
-// import Particles, { initParticlesEngine } from '@tsparticles/react';
+import particlesConfig from '../../particles.config';
+import { useEffect, useState } from 'react';
+import Particles, { initParticlesEngine } from '@tsparticles/react';
+import { loadSlim } from '@tsparticles/slim';
 
-// import { loadSlim } from '@tsparticles/slim';
-// const ParticlesBackground = () => {
-//   const [init, setInit] = useState(false);
-//   useEffect(() => {
-//     initParticlesEngine(async (engine) => {
-//       await loadSlim(engine);
-//     }).then(() => {
-//       setInit(true);
-//     });
-//   }, []);
-//   const particlesLoaded = (container) => {
-//     console.log(container);
-//   };
+const ParticlesBackground = () => {
+  const [init, setInit] = useState(false);
+  useEffect(() => {
+    initParticlesEngine(async (engine) => {
+      await loadSlim(engine);
+    }).then(() => {
+      setInit(true);
+    });
+  }, []);
+  const particlesLoaded = (container) => {
+    console.log(container);
+  };
 
-//   return (
-//     <div>
-//       <Particles
-//         id="tsparticles"
-//         particlesLoaded={particlesLoaded}
-//         options={particlesConfig}
-//       />
-//     </div>
-//   );
-// };
+  return (
+    <>
+      {init && (
+        <Particles
+          id="tsparticles"
+          particlesLoaded={particlesLoaded}
+          // the particlesconfig.js contains all the json required for particleJS
+          options={particlesConfig}
+        />
+      )}
+    </>
+  );
+};
 
-// export default ParticlesBackground;
+export default ParticlesBackground;
